+++ resources/release-notes.md
@@ -1,3 +1,7 @@
+## App
+
+- (IMAP) Allow specifying data level for processing (L1a or L1b)
+
 ## Software
 
 - (All) Add ability to set figure resolution via MATLAB settings: